+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `From<Number>`, `From<usize>` and `From<isize>` for `Value`.
 - Add `Default` for `Value`.
 - Add `PartialEq` for `ValueRef`.
+- Add `pointer` for `Value` and `ValueRef`.
 - Add `json!` macro. 
 
 ## [0.1.2] - 2023-10-30